# docs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: DPO оператора
  description: "REST API центра управления согласиями субъектов персональных данны\
    х. Идентификаторы — на английском, пользовательские тексты — на русском (NFR-8)."
  license:
    name: Proprietary
  title: In consensu — центр управления согласиями
  version: v1
servers:
- description: Текущий хост
  url: /
paths:
...
  /api/v1/subjects/{id}/transfers:
    get:
      operationId: transfersOf
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferView"
          description: OK
      tags:
      - transfer-controller
...
  /api/v1/transfers/check:
    post:
      operationId: check
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferCheckRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferCheckResponse"
          description: OK
      tags:
      - transfer-controller
...
components:
  schemas:
...
    ThirdPartyRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
...
    TransferCheckRequest:
      type: object
      properties:
        pdnCategories:
          type: array
          items:
            type: string
        subjectId:
          type: string
          format: uuid
        thirdPartyId:
          type: string
          format: uuid
      required:
      - subjectId
      - thirdPartyId
    TransferCheckResponse:
      type: object
      properties:
        allowed:
          type: boolean
        allowedCategories:
          type: array
          items:
            type: string
        deniedCategories:
          type: array
          items:
            type: string
        reason:
          type: string
        validUntil:
          type: string
          format: date-time
    TransferView:
      type: object
      properties:
        basisConsentId:
          type: string
          format: uuid
        daysLeft:
          type: integer
          format: int64
        pdnCategories:
          type: array
          items:
            type: string
        thirdParty:
          $ref: "#/components/schemas/ThirdPartyRef"
        thirdPartyContractExpired:
          type: boolean
        validUntil:
          type: string
          format: date-time
...